+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.1 [unreleased]
 
+- Added support for many more types
+- Fixed output order with `output_names` option
 - Fixed `File doesn't exist` on Windows
 
 ## 0.1.0