+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 [unreleased]
 
+- Packaged ONNX Runtime with gem
 - Added support for many more types
 - Fixed output order with `output_names` option
 - Fixed `File doesn't exist` on Windows